requirements.txt
streamlit
pandas
requests
matplotlib
numpy
streamlit_echarts
pyecharts 
python-dotenv
google-generativeai
PyPDF2 
pdfplumber
openai
langchain
langchain-community 
langchain-core
pypdf
tiktoken
faiss-cpu
transformers
openai
langchain
langchain-community 
langchain-core
pypdf
beautifulsoup4
langchain_openai
chromadb
